Bump crate versions and fix SQLite count types
Update Cargo.lock to bump git-meta-cli and git-meta-lib from 0.1.6 to
0.1.7. Fix several SQLite query result type usages by reading COUNT(*) as
i64 and converting to u64, and normalize query_row type handling to avoid
ambiguous type inference. These changes prevent incorrect integer type
assumptions and ensure counts are safely converted to the expected u64
types.

+++ crates/git-meta-lib/src/db/promised.rs
@@ -95,7 +95,7 @@ impl Store {
             "SELECT target_type, COUNT(*) FROM metadata WHERE is_promised = 1 GROUP BY target_type ORDER BY target_type",
         )?;
         let rows = stmt.query_map([], |row| {
-            Ok((row.get::<_, String>(0)?, row.get::<_, u64>(1)?))
+            Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)? as u64))
         })?;
         let mut results = Vec::new();
         for row in rows {
